Refactor: Уточняет типы данных в моделях
*   В модели `Component_usage`:
    *   Изменяет тип поля `project_id` с `int` на `long`.
    *   Изменяет тип поля `component_id` с `int` на `long`.
    *   Обновляет сигнатуры конструктора, геттеров и сеттеров для `project_id` и `component_id` в соответствии с новым типом `long`.
*   В модели `Component`:
    *   Изменяет тип поля `price` с `int` на `float`.
    *   Обновляет сигнатуры конструктора, геттера и сеттера для `price` в соответствии с новым типом `float`.

+++ src/main/java/com/monolatte/kontur/model/Notes/Component.java
@@ -6,9 +6,9 @@ public class Component {
     String type;
     String specification;
     String datasheet_link;
-    int price;
+    float price;
 
-    public Component(String name, String type, String specification, String datasheet_link, int price) {
+    public Component(String name, String type, String specification, String datasheet_link, float price) {
         this.name = name;
         this.type = type;
         this.specification = specification;
@@ -59,11 +59,11 @@ public class Component {
         this.datasheet_link = datasheet_link;
     }
 
-    public int getPrice() {
+    public float getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(float price) {
         this.price = price;
     }
 }
